+++ src/test/functional/java/gov/ca/cwds/api/authorize/ClientAccessForStaffInLakeCountyWithSensitiveAccess.java
@@ -1,21 +1,31 @@
-package gov.ca.cwds.api.client.access;
+package gov.ca.cwds.api.authorize;
+
+import static io.restassured.RestAssured.given;
+import static org.junit.Assert.assertFalse;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
-import static io.restassured.RestAssured.given;
-import static org.junit.Assert.assertFalse;
-import java.util.HashMap;
-import java.util.Map;
+
 import gov.ca.cwds.api.FunctionalTest;
 import gov.ca.cwds.api.builder.HttpRequestHandler;
 import gov.ca.cwds.rest.authenticate.UserGroup;
 import gov.ca.cwds.rest.core.Api;
 import io.restassured.http.ContentType;
-import io.restassured.response.Response;
 
-public class ClientAccessForStaffInStateWithSensitiveAccess extends FunctionalTest {
+/**
+ * 
+ * Functional test for client access (/authorize) when the staff person (user)
+ * has sensitive access and is in Lake county.
+ * 
+ * Reference Section 4 of the Security_V2.7 document (Table 4 -  Client Search Access).
+ * 
+ * @author CWDS API Team
+ *
+ */
+public class ClientAccessForStaffInLakeCountyWithSensitiveAccess extends FunctionalTest {
   String resourcePath;
   private HttpRequestHandler httpRequestHandler;
+  
 
   /**
    * 
@@ -23,30 +33,24 @@ public class ClientAccessForStaffInStateWithSensitiveAccess extends FunctionalTe
   @Before
   public void setup() {
     // logged in staff with Sensitive access and
-    // USERID->STAFF_PERSON->CWS_OFFICE.Government_Entity_type=1126 (California)
+    // USERID->STAFF_PERSON->CWS_OFFICE.Government_Entity_type=1084 (Lake)
     resourcePath = getResourceUrlFor("/" + Api.RESOURCE_AUTHORIZE + "/client" + "/{id}");
     httpRequestHandler = new HttpRequestHandler();
-    this.loginUserGroup(UserGroup.STATE_SENSITIVE);
+    this.loginUserGroup(UserGroup.COUNTY_SENSITIVE);
+
   }
 
-  @Test 
-  public void testInvalidClientId() {
-    given().pathParam("id", "999999XYZ").queryParam(httpRequestHandler.TOKEN, token)
-    .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
-    .statusCode(200);
-  }  
-  
   @Test
   public void shouldReturnClientWithNoAccessRestrictions() {
     given().pathParam("id", "CFOmFrm057").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
+    
   }
   
   @Test
-  public void shouldReturnClientInSameCountyWithSensitive() {
-    // requires sensitive client owned by 1126 (California)
-    given().pathParam("id", "1S3k0iH00T").queryParam(httpRequestHandler.TOKEN, token)
+  public void shouldReturnClientInSameCountyWithSensitive() {    this.loginUserGroup(UserGroup.COUNTY_SENSITIVE);
+    given().pathParam("id", "B5mi8Qr00T").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
     
@@ -54,20 +58,18 @@ public class ClientAccessForStaffInStateWithSensitiveAccess extends FunctionalTe
   
   @Test
   public void shouldNotReturnClientInSameCountyWithSealed() {
-    // requires sealed client owned by 1126 (California)
    given().pathParam("id", "4kgIiDy00T").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(403);
-     
+    
   }
  
   @Test
   public void shouldNotReturnClientInDifferentCountyWithSensitive() {
-    // should return status 403
-    given().pathParam("id", "9PIxHucCON").queryParam(httpRequestHandler.TOKEN, token)
+   given().pathParam("id", "9PIxHucCON").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
-    .statusCode(200);
-    
+    .statusCode(403);
+   
   }
   
   @Test
@@ -79,13 +81,12 @@ public class ClientAccessForStaffInStateWithSensitiveAccess extends FunctionalTe
   }
   
   @Test
-  public void shouldNotReturnClientInNoCountyWithSensitive() {
+  public void shouldReturnClientInNoCountyWithSensitive() {
     // client with limited access code = 'S' and government entity of 1126 (California)
-    // should not be able to access client - return status 403
     given().pathParam("id", "AYk7k55aaf").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
-    
+  
   }
   
   @Test
@@ -94,7 +95,7 @@ public class ClientAccessForStaffInStateWithSensitiveAccess extends FunctionalTe
     given().pathParam("id", "BK3EnRK0DE").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(403);
-    
+   
   }
 
 }
