+++ src/test/functional/java/gov/ca/cwds/api/authorize/ClientAccessForStaffInStateWithSealedAccess.java
@@ -1,17 +1,14 @@
-package gov.ca.cwds.api.client.access;
-
+package gov.ca.cwds.api.authorize;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import static io.restassured.RestAssured.given;
-import static org.junit.Assert.assertFalse;
 import gov.ca.cwds.api.FunctionalTest;
 import gov.ca.cwds.api.builder.HttpRequestHandler;
 import gov.ca.cwds.rest.authenticate.UserGroup;
 import gov.ca.cwds.rest.core.Api;
 import io.restassured.http.ContentType;
 
-public class ClientAccessForStaffInLakeCountyWithSealedAccess extends FunctionalTest {
+public class ClientAccessForStaffInStateWithSealedAccess extends FunctionalTest {
   String resourcePath;
   private HttpRequestHandler httpRequestHandler;
 
@@ -20,11 +17,11 @@ public class ClientAccessForStaffInLakeCountyWithSealedAccess extends Functional
    */
   @Before
   public void setup() {
-    // logged in staff with Sensitive access and
-    // USERID->STAFF_PERSON->CWS_OFFICE.Government_Entity_type=1084 (Lake)
+    // logged in staff with Sealed access and
+    // USERID->STAFF_PERSON->CWS_OFFICE.Government_Entity_type=1126 (California)
     resourcePath = getResourceUrlFor("/" + Api.RESOURCE_AUTHORIZE + "/client" + "/{id}");
     httpRequestHandler = new HttpRequestHandler();
-    this.loginUserGroup(UserGroup.COUNTY_SEALED);
+    this.loginUserGroup(UserGroup.STATE_SEALED);
   }
 
   @Test
@@ -33,23 +30,27 @@ public class ClientAccessForStaffInLakeCountyWithSealedAccess extends Functional
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
   }
-  
+   
   @Test
   public void shouldNotReturnClientInSameCountyWithSensitive() {
-    given().pathParam("id", "B5mi8Qr00T").queryParam(httpRequestHandler.TOKEN, token)
-    .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
-    .statusCode(403);
-    
-  }
-  
-  @Test
-  public void shouldReturnClientInSameCountyWithSealed() {
-    given().pathParam("id", "Ba29OOP75a").queryParam(httpRequestHandler.TOKEN, token)
+    // requires sealed client owned by county 1126 (California)
+    // should return status 403
+    given().pathParam("id", "1S3k0iH00T").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
     
   }
- 
+  
+  @Test
+  // requires sealed client owned by county 1126 (California)
+  // should return status 200
+  public void shouldReturnClientInSameCountyWithSealed() {
+    given().pathParam("id", "4kgIiDy00T").queryParam(httpRequestHandler.TOKEN, token)
+    .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
+    .statusCode(403);
+     
+  }
+
   @Test
   public void shouldNotReturnClientInDifferentCountyWithSensitive() {
     given().pathParam("id", "9PIxHucCON").queryParam(httpRequestHandler.TOKEN, token)
@@ -68,8 +69,7 @@ public class ClientAccessForStaffInLakeCountyWithSealedAccess extends Functional
   
   @Test
   public void shouldNotReturnClientInNoCountyWithSensitive() {
-    // client with limited access code = 'S' and government entity of 1126 (California)
-    // should not be able to attach/use sensitive client - status 403
+    // should return status 403 - should not be able to attach client with sensitive access
     given().pathParam("id", "AYk7k55aaf").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
@@ -78,7 +78,6 @@ public class ClientAccessForStaffInLakeCountyWithSealedAccess extends Functional
   
   @Test
   public void shouldReturnClientInNoCountyWithSealed() {
-    // client with limited access code = 'R' and government entity of 1126 (California)
     given().pathParam("id", "BK3EnRK0DE").queryParam(httpRequestHandler.TOKEN, token)
     .contentType(ContentType.JSON).accept(ContentType.JSON).when().get(resourcePath).then()
     .statusCode(200);
